Moved files into subfolders streams and resources. Changed res_ref operator const char * into more explicit CString() method. Changed GFX::rect w and h members to uint16 from sint16. Various other changes.

+++ Bitmap.cpp
@@ -49,7 +49,7 @@ Bitmap::SetPalette(const Palette& palette)
 
 
 void
-Bitmap::SetColorKey(uint8 index)
+Bitmap::SetColorKey(uint32 index)
 {
 	SDL_SetColorKey(fSurface, SDL_SRCCOLORKEY|SDL_RLEACCEL, index);
 }
